Resolves sign_in exception.
It resolves a bug related to the user being signed in and remembered. After session expiration, when the user wants to use the app, the system logs the user out and tries to log the change (audited gem). Since the gem is configured to log remember_token, and remember_created_at fields it tries to fetch current user and the whole process starts again, creating infinite loop and producing stack level too deep exception.
The fix is to disable logging changes for remember_token and remember_created_at for audited gem.